+++ spectrafit/report.py
@@ -3,7 +3,7 @@ import pprint
 
 from typing import Any
 from typing import Dict
-from typing import List
+from typing import Hashable
 from typing import Optional
 from typing import Tuple
 
@@ -127,8 +127,12 @@ class RegressionMetrics:
             else (np.array([_true]), np.array([_pred]))
         )
 
-    def __call__(self) -> Dict[str, Any]:
-        """Calculate the regression metrics of the Fit(s) for the post analysis."""
+    def __call__(self) -> Dict[Hashable, Any]:
+        """Calculate the regression metrics of the Fit(s) for the post analysis.
+
+        Returns:
+            Dict[Hashable, Any]: Dictionary containing the regression metrics.
+        """
         metrics_fnc = (
             explained_variance_score,
             r2_score,
@@ -140,7 +144,7 @@ class RegressionMetrics:
             mean_absolute_percentage_error,
             mean_poisson_deviance,
         )
-        metric_dict: Dict[str, List[float]] = {}
+        metric_dict: Dict[Hashable, Any] = {}
         for fnc in metrics_fnc:
             metric_dict[fnc.__name__] = []
             for y_true, y_pred in zip(self.y_true.T, self.y_pred.T):
@@ -148,7 +152,7 @@ class RegressionMetrics:
                     metric_dict[fnc.__name__].append(np.nan)
                 else:
                     metric_dict[fnc.__name__].append(fnc(y_true, y_pred))
-        return metric_dict
+        return pd.DataFrame(metric_dict).T.to_dict(orient="split")
 
 
 def fit_report_as_dict(
@@ -305,10 +309,10 @@ class PrintingResults:
         """
         print(
             tabulate(
-                pd.DataFrame.from_dict(args).T,
+                pd.DataFrame(**args).T,
                 headers="keys",
                 tablefmt="fancy_grid",
-                floatfmt=".2f",
+                floatfmt=".3f",
             )
         )
 
@@ -332,7 +336,7 @@ class PrintingResults:
                 )
             except MinimizerException as exc:
                 print(f"Error: {exc} -> No confidence interval could be calculated!")
-                self.args["confidence_interval"] = None
+                self.args["confidence_interval"] = {}
         print("\nOverall Linear-Correlation:\n")
         self.print_tabulate(args=self.args["linear_correlation"])
         print("\nRegression Metrics:\n")