+++ spectrafit/utilities/__init__.py
@@ -1,0 +1,1 @@
+"""Utilities of the spectraFit package."""